MDL-72232 js: Add ability to set a modal to be vertically centered
With this change the modal can be configured to be vertically centered
on a page or not. This can be done by setting the modal configuration
option 'verticallyCentered' to true or false when creating a modal
instance. If 'verticallyCentered' is not explicitely configured,
it will be set to 'false' by default.