MDL-XXXXX behat: Check if the toggle action element is visible
Checks whether the toggle action element is visible before executing
the call action to toggle the administration menu. This check is
necessary as this is also used to toggle the more dropdown menu in the
secondary navigation menu which in some cases can be hidden when there
aren't any items that should be placed under the more menu.