MDL-74740 boost: Display the skip links from the blocks drawer
Due to the existing 'z-index' css property on the drawer element, the
skip links from the drawer were stacked behind the navbar. To solve
this problem the 'z-index' property will not be added to the drawer
element when the current focus is on an element within the drawer. In
any other case, the original stacking order is preserved. In addition
to this change, the top position and the height of the drawers is
being fixed.