MDL-79531 mod_quiz: Fix exception when adding random questions
When adding a random question, an exception is thrown in the console if
the selected category does not contain any questions. This happens
because the code is trying to obtain the value of a data property
(data-filtercondition) of an element that does not exist (filter
condition element). This problem also occurs when creating a random
question in a new category although this data attribute is not required
to create the new category.
This change is making sure that the element exists (which means the
selected category has questions) before creating the random question.
The value of the data-filtercondition attribute is no longer fetched
when the user is attempting to create a random question in a new
category.