MDL-80741 mod_assign: Ensure multiple attempts are properly set
A recent change in the default option for "maxattempts" no longer
permits unlimited attempts by default. As a result, it is necessary
to update all instances in the code where multiple attempts are
expected. This is achieved by explicitly setting the "maxattempts"
option to either unlimited (-1) or a value greater than 1.