MDL-81714 tool_recyclebin: Remove unnecessary task runs in tests
I assume that it was necessary at some point to run the ad-hoc tasks in
these tests, however the only task now being run is
regrade_final_grades, which causes the tests to fail due as they produce
output.

Whether the regrade is performed or not has no impact on the result of
the test, so removing the ad-hoc task run seems appropriate.