MDL-85331 mod_lti: Update instance creation logic in data generator
Updates the create_instance() method in the LTI module data generator
class to align with the new prerequisites set by the resource link
manager. Based on this requirement we need to ensure the type ID of
an existing tool is passed before creating the module, with an
enabled mod_lti:activityplacement placement type associated.