MDL-84653 core_ltix: Relax validation in resource link manager
Modifies the resource link manager class to remove validation that
was enforced during every initialization. This validation is only
useful when creating a resource link to ensure the correctness of
the data being added. In other cases such as retreiving, updating or
deleting resource links - it is redundant and sometimes restrictive.
The validation logic has therefore been moved to execute only during
resource link creation. Additionally, the factory method create()
has been removed as it is unnecessary, and all methods associted to
CRUD operations are now static. All usages of the resource link
manager and tests have been updated accordingly.